Ignore migration errors when deleting a cgroup
As libcg tries to migrate processes before deleting a cgroup,
it can encounter a race-condition if a group process was killed
just before calling. To avoid generating errors even when everything
went fine, we ask the libcg to ignore migration errors and to continue
deleting the cgroup.
This should make disappear a number of spurious warnings on system
cleanup.